AG-7679 - Extract InteractionManager event registration code into BaseManager.
